Refresh auth token if expired, using /refresh route.
We now store both auth token and refresh token in local storage. If auth token is expired, we refresh it by contacting the server using the refresh token. If the refresh token is expired, the user gets logged out. We check if the user is logged in if they have a refresh token with the Materials API, removing the need to store the userInfo information in local storage.